MOOSE Property Section now uses real checkboxes
The MOOSE Property Section has been changed so that, instead of using a
content provider to put pictures of check boxes in non-selected cells,
the check box column is created by the table, which is now a
CheckboxTableViewer. This should fix issues with the view not displaying
correctly under gtk. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>